LibDebug: Move get_attribute_value to the DwarfInfo class
