+++ Userland/Libraries/LibDebug/Dwarf/DwarfInfo.h
@@ -16,6 +16,32 @@
 
 namespace Debug::Dwarf {
 
+struct AttributeValue {
+    enum class Type : u8 {
+        UnsignedNumber,
+        SignedNumber,
+        LongUnsignedNumber,
+        String,
+        DieReference, // Reference to another DIE in the same compilation unit
+        Boolean,
+        DwarfExpression,
+        SecOffset,
+        RawBytes,
+    } type;
+
+    union {
+        u32 as_u32;
+        i32 as_i32;
+        u64 as_u64;
+        const char* as_string; // points to bytes in the memory mapped elf image
+        bool as_bool;
+        struct {
+            u32 length;
+            const u8* bytes; // points to bytes in the memory mapped elf image
+        } as_raw_bytes;
+    } data {};
+};
+
 class DwarfInfo {
 public:
     explicit DwarfInfo(const ELF::Image&);
@@ -26,6 +52,9 @@ public:
 
     template<typename Callback>
     void for_each_compilation_unit(Callback) const;
+
+    AttributeValue get_attribute_value(AttributeDataForm form,
+        InputMemoryStream& debug_info_stream, const CompilationUnit* unit = nullptr) const;
 
 private:
     void populate_compilation_units();
